check CRC of input packets in CLI tool
This checks the CRC-32 of each input packet in the CLI and prints
a warning if the CRC is wrong. One reason why this is helpful is
because it makes it more obvious if the user is feeding in SSDV
packets with the wrong format into the CLI tool.

The SSDV packets are processed regardless of whether their CRC is
correct, however. For encoding, they are taken into account, while
for decoding they will be discarded by the FEC decoder.

Signed-off-by: Daniel Estévez <daniel@destevez.net>