add support for other SSDV packet formats
This adds some traits that can be used to define different SSDV
packet formats. The standard no-FEC packet format and the Longjiang-2
packet format are now implemented using these traits. The command
line tool gains a --format parameter to choose the format. It defaults
to the no-FEC format, since that is the format used by the upstream
ssdv encoder/decoder software.

Signed-off-by: Daniel Estévez <daniel@destevez.net>